Added setting team across the app
Eventually this will work with a multiorg approach, but right now
it works by explicitly setting the team via the Auth Control
screen. This may be hardcoded in the future; URLs will at least
need to adapt to eventually supporting multiple teams.

I may do more work on this before submitting it; right now the code
is a mess and takes three different approaches, because I wasn't
sure what the right approach was. I think I have settled on a good
format for everything by the end, but I need to go back and fix
some in progress work.

Hopefully this can be rolled out as "good enough" to start working
on other features.