Rename some items from Owner to User
Ideally, the User of a Member will refer to the Django User object.
This should be made consistent on the backend. Currently, there is
an 'owner' of a member, which feels a bit ambiguous/misleading.
However, until this is changed on the backend, it will remain this
way on the frontend.