Added userNameForUser and test, plus added to navbar
Navbar should now include the username. This will either give the
first and last name, falling back to first name if last name is
undefined or blank, falling back to username if first name is
undefined or blank.

Tests have been added to ensure the above functionality.

There is an ESLint change that needed to be made to make this work.
There's some conflict with camelcase vs. the record titles. This
must have cropped up with an error in vscode workspaces differing.
Going to work on getting this standardized so this doesn't occur
again.