Removed async setup and migrated to lifecycle hooks.
This commit will fix some of the errors of the Vue SPA effectively
failing to render and crashing after switching between the Teams
screens and other screens (such as Help and Auth).

Accidentally used an experimental feature to implement calling load
from startup. Should have been obviously wrong but wasn't to me.

If anyone encounters a similar need, make sure to use the lifecycle
hooks to trigger API calls (such as loading before the component is
mounted). This is the intended way to do things and won't run into
errors.